Search related files as default for :Efoo
Adds a fallback for the navigation commands to search related files for a
file matching the corresponding type. Related files are searched
recursively, breadth-first, and recursion is limited to a maximum depth.
In addition to moving "outward" from the current file, the search moves
"inward" from files matching the target type. If the "inward" search for
a file intersects with the "outward" search, that file is considered a
match for navigation.

Fixes #80, #26, #22