Limit BufFilePost detection to known good buffers
This is targeted explicitly at terminal buffers, which obnoxiously
trigger BufFilePost before 'buftype' is set.  We could apply crude
heuristics like looking for the "!" at the beginning of the buffer name,
but a more robust and general solution is to ignore any buffer that some
other event didn't trigger on first.

Closes https://github.com/tpope/vim-projectionist/issues/155