Provide mechanism to opt out of detection
This is my workaround to the problem of editing files on network shares
being slow due to the large number of synchronous IO calls Projectionist
can end up doing during detection.  It can also work around other
problems, such as the annoying GUI notifications Keybase fires off when
a Projectionist check happens to match a Keybase user directory.

I've kept the implementation and feature set minimal so that other
plugins can replicate it with little effort.  This means
ExcludeBufferFromDiscovery() should be liberal in what it accepts and
handle making the path absolute, normalizing slashes on Windows, etc.
Another plugin could step in here to act as an intermediary and do the
heavy lifting, and I might make one of those some day, but I currently
have no plans to do so.

This is what I'm currently using:

        let s:slash = exists('+shellslash') ? '\' : '/'
        function! ExcludeBufferFromDiscovery(file, ...) abort
          let file = tr(a:file, s:slash, '/')
          if file !~# '^\a\+:\|^/'
            let file = tr(getcwd(), s:slash, '/') . '/' . file
          endif
          if has('win32')
            return file =~# '^//'
          else
            return file =~# '^/\%(run\|media\|Volumes\)/'
          endif
        endfunction