Fix {dirname} for top level file
If dirname() returns an empty string, then "{dirname}/foo" will resolve
to the absolute path "/foo" instead of a relative path.  The natural fix
is to return ".", but we need to take care to normalize it so that it
doesn't get treated as "current working directory".

References: https://github.com/tpope/vim-projectionist/pull/190