Support filename(line,col) like filename:line:col
The format of "filename:line:col", with an optional col, is very common
in the output of UNIX commands (e.g., grep).  I made Projectionist DWIM
for a filename arguments like this so that I could use :A with a pasted
argument as an easy way to jump to such a notated position.

Similarly, the format of "filename(line,col)" seems to be common in the
output of Microsoft commands.  I'm less familiar with that world, but
lately I've been encountering the format a lot in the TypeScript
ecosystem.  It seems reasonable (and convenient to my needs) to give it
the same treatment as the UNIX format.

I've allowed an optional trailing colon in both formats, since
double/triple clicking to select text in a terminal oftentimes slurps up
the subsequent colon as well.